+++ application/classes/Core/Theme.php
@@ -1,5 +1,6 @@
 <?php
 namespace Core {
+	use Registry;
 
 	/**
 	* Loads and setups the current configured theme in use.
@@ -16,7 +17,7 @@ namespace Core {
 		* Doesn't do much of interest, this shouldn't be required to mess with.
 		* @return void
 		*/
-		public function __construct($themename) {
+		public function __construct(string $themename) {
 			$this->name = $themename;
 			$this->theme = (new Configuration($this->getTemplatePath("/theme.json")));
 
@@ -27,7 +28,7 @@ namespace Core {
 		* Get the path to the current active theme.
 		* @return string
 		*/
-		public function getTemplatePath($tpl = '') {
+		public function getTemplatePath(string $tpl = '') : string {
 			$path = \Registry::get("Core\Application")->getApplicationPath()."/application/themes/".$this->getName();
 
 			if($tpl != '') {
@@ -44,7 +45,7 @@ namespace Core {
 		* @author Allan Thue Rehhoff
 		* @return string
 		*/
-		public function getDirectoryUri($file = '/') {
+		public function getDirectoryUri(string $file = '/') : string {
 			$protocol = SSL ? "https://" : "http://";
 			$host  = isset($_SERVER["SERVER_NAME"]) ? $_SERVER["SERVER_NAME"] : "127.0.0.1";
 			$path = rtrim(dirname($_SERVER["PHP_SELF"]), "/\\");
@@ -57,7 +58,7 @@ namespace Core {
 		* Get the current theme name loaded.
 		* @return string
 		*/
-		public function getName() {
+		public function getName() : string {
 			return $this->name;
 		}
 
@@ -68,16 +69,19 @@ namespace Core {
 		* @uses \Document
 		* @return void
 		*/
-		private function addAssets() {
+		private function addAssets() : void {
 			if(!empty($this->theme->get("javascript"))) {
 				foreach($this->theme->get("javascript") as $javascript) {
 					if(filter_var($javascript, FILTER_VALIDATE_URL)) {
 						$src = $javascript;
 					} else {
+						// Do not add files that does not exist in theme
+						if(file_exists($this->getTemplatePath($javascript)) === false) continue;
+
 						$src = $this->getDirectoryUri($javascript);
 					}
 
-					\Registry::get("Document")->addJavascript($src);
+					Registry::get("Document")->addJavascript($src);
 				}
 			}
 
@@ -86,10 +90,13 @@ namespace Core {
 					if(filter_var($stylesheet, FILTER_VALIDATE_URL)) {
 						$src = $stylesheet;
 					} else {
+						// Do not add files that does not exist in theme
+						if(file_exists($this->getTemplatePath($stylesheet)) === false) continue;
+
 						$src = $this->getDirectoryUri($stylesheet);
 					}
 
-					\Registry::get("Document")->addStylesheet($src);
+					Registry::get("Document")->addStylesheet($src);
 				}
 			}
 		}